--- src/css/list.scss
@@ -1,24 +1,0 @@
-@import 'reset-css';
-
-.list {
-  display: flex;
-  flex-wrap: wrap;
-  margin: auto;
-  width: 1000px;
-
-  &__item {
-    display: flex;
-    margin: 0 1%;
-    width: 31.33333%;
-
-    &:nth-child(n+4) {
-      margin-top: 15px;
-    }
-  }
-
-  &__thumb {
-    > img {
-      max-width: 100%;
-    }
-  }
-}
